POC of Propert Based Tests
I should be able to use FsCheck to do property based testing on anything
that isn't dependent on Fable. This setup proves that correct.

To do Property Based testing on the game loop I'll need to modify the
render dependency, which currently has a dependency on an HTML image. If
I do that then I can property test everything except the renderer, and
I'm okay with that.