WIP: Create a project to run FSCheck tests on game
This is a WIP because in order to get it working I had to create a new project using Forge, which put things in a different direction structure. Therefore everything is all out of place.

That said propertyTests/PropertyTests is capable of running FSCheck tests from Fake and it works.