Simpilfy the Presentation layer
Rather than having a lookup table into a constant, which in turn goes into
another lookup table, have the presentation layer take the lookup table from
entity to 'HTMLImage, where 'HTMLImage is a template. This not only makes things
a lot easier to follow and simplifies the testing, but I believe it removes a
dependency on Fable in the presentation layer. That might make it better served
by property based testing.