Huge Reorg of the types and modules
Well this may not be as big a reorg as planned - I had a lot of trouble
getting this to compile again. Ultimately the game modules are under the
namespace "SpaceInvaders", the engine modules are under the namespace
"Engine" and the tests are all in the namespace "Test". In order to make
sure that I don't get naming conflicts the files in UnitTest are no
longer just able to be namespce Test, but need to be module
Test.<ModuleName>. I'm not 100% sure how this ever woked but I need to
get a commit in, and the tests pass.