Make the Game have explicit things
This is a refactoring step, I'm restoring the idea that the game has a Laser,
Invaders, Bullets. This will make a lot of the functions remove their match
statements eventually. In fact the union for EntityProperties may no longer be
needed.

That said currently the laser is in the Entities list and in Laser/option. The
next step is to use the Laser field, removing it from the entities.