Start moving to sending entity,properties tuples
This passes the tests and compiles, so even though it's not a complete finish
and the property tests fail, I want to get a commit in.

I'm trying to solve my infamous typing problem by making the modules take an
entity,property tuple. This means that you'll do the switch statement (to
destructure) before you call the helper functions.