Remove fromEntity, move back to types
Not sure what I was thinking with the tuple thing. The problem with a
BulletEntity was that isOverlapping assumes entities. However if we make the
entire Entity module Location, everything is fine.

To that end I've made some of Entity take locations, and removed the `fromEntity`
function from the Laser. `toEntity` is still present because we need it for
presentation. This looks like the way to go.